pyproject.toml
[project]
name = "store-douyin"
version = "0.1.0"
description = "抖音自动化 Skills，基于 CDP 浏览器自动化"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "setuptools>=65.0",
    "websockets>=12.0",
    "requests>=2.28.0",
    "yt-dlp>=2024.1.0",
    "openai-whisper>=20231117",
    "soundfile>=0.12.0",
    "scipy>=1.10.0",
    "openai>=1.0.0",
    "Pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "numpy<2",
    "imageio-ffmpeg>=0.5.0",
    "moviepy>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
    "pytest>=8.0",
]
material = [
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "onnxruntime>=1.19.2",
    "torch>=2.4.0",
    "protobuf<=3.20.3",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
]
ignore = [
    "E402",   # module-level imports not at top
    "RUF001", # ambiguous unicode characters
    "RUF002", # ambiguous unicode in docstrings
    "RUF003", # ambiguous unicode in comments
]

[tool.ruff.lint.isort]
known-first-party = ["douyin"]

[tool.pytest.ini_options]
testpaths = ["tests"]
